+++ js/app/controllers/calendarlistcontroller.js
@@ -145,7 +145,7 @@ app.controller('CalendarListController', ['$scope', '$rootScope', '$window', 'Ha
 
 		$scope.sendMail = function (item) {
 			item.toggleSendingMail();
-			MailerService.sendMail(item.email, item.publicSharingURL, item.calendar.displayname).then(function (response) {
+			MailerService.sendMail(item.email, item.calendar.publicToken, item.calendar.displayname).then(function (response) {
 				if (response.status === 200) {
 					item.email = '';
 					OC.Notification.showTemporary(t('calendar', 'Email has been sent.'));