+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased] - XXXX-XX-XX
+
+
 ## [2.0.0] - 2022-07-05
 
 ### Changed